Split larger code out.
