+++ Basis/Packages/com.basis.framework/BasisUI/PanelComponents/PanelSectionDividerManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2f81f6c5e2e84cf6a4f77259df2c9a13
